Support permalinks
This is done by overtaking the rendering for h2 and h3 tags, parsing their content and generating an id using that, inserting an <a> link with that id. The link is styled so that it shows up when hovering over the heading, to be minimally intrusive.

Also tweaked the styling of external vs internal links to highlight internal links with the accent color a bit more and to reduce the number of underlines.