Generate og images for blog posts
Spent way too much time on this, but it's working now. It uses the
excellent https://github.com/vercel/satori package to render the svg for
the preview, which is then converted to png. Thanks vercel & wasm!

The canonical urls were incorrect because of the ad hoc way were adding
`blog/` to various places, so refactored that to all flow through the
main function. As part of this, the hostname is now an input to the
script, and locally we pass in localhost as an input on purpose to make
sure it works correctly.